starting to refactor code
